requirements.txt
streamlit>=1.0.0
Pillow>=9.0.0
requests>=2.28.0
volcengine-python-sdk[ark]>=1.0.0
python-dotenv>=1.0.0
supabase>=2.0.0

